+++ R/domain_report.R
@@ -32,3 +32,5 @@ domain_report <- function(domain = NULL, ...) {
 }
 
 
+
+