+++ NEWS.md
@@ -1,3 +1,8 @@
+# virustotal 0.1.1
+
+* Enforces rate limiting --- 4 queries per minute. 
+* Graceful error handling if error limit exceeded.
+
 # virustotal 0.1.0
 
 * Initial release